+++ init_trap/config/config.go
