+++ init_trap/main_test.go
